feat(variable-explanation): Showing coordinates of mouse for changing root coordinates of stack pane.

+++ src/main/java/de/uniks/stp24/model/Effect.java
@@ -2,8 +2,8 @@ package de.uniks.stp24.model;
 
 public record Effect(
         String variable,
-        int base,
-        int multiplier,
-        int bonus
+        double base,
+        double multiplier,
+        double bonus
 ) {
 }
